refact: rename routes directory to router; implement routing logic and authentication protection; update path aliases in tsconfig
