+++ src/index.tsx
@@ -3,8 +3,8 @@ import { createRoot } from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 import { Provider } from "react-redux";
 import { store } from "@/store";
-import { browserRouter } from "@/routes";
 import "./index.css";
+import { browserRouter } from "./router";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
