+++ src/router/logic/AuthProtectedRouteLogic.tsx
@@ -4,10 +4,10 @@ interface ProtectedRouteProps {
   children: React.ReactNode;
 }
 
-const ProtectedRoute: React.FC<ProtectedRouteProps> = ({ children }) => {
+const AuthProtectedRouteLogic: React.FC<ProtectedRouteProps> = ({ children }) => {
   // Add your authentication logic here
   // For now, we'll just return the children
   return <>{children}</>;
 };
 
-export default ProtectedRoute;
+export default AuthProtectedRouteLogic;
